requirements.txt
asgiref==3.7.2
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4